+++ changelog.md
@@ -6,6 +6,7 @@
 * [1395](https://github.com/zeta-chain/node/pull/1395) - Add state variable to track aborted zeta amount
 * [1387](https://github.com/zeta-chain/node/pull/1387) - Add HSM capability for zetaclient hot key
 * enable zetaclients to use dynamic gas price on zetachain - enables >0 min_gas_price in feemarket module
+* add static chain data for Sepolia testnet
 
 ### Fixes
 * fix go-staticcheck warnings for zetaclient